### 2025-02-10 #684 @rmgpanw
gtexr: Query the GTEx Portal API
Submitting Author Name: Alasdair Warwick
Submitting Author Github Handle: @rmgpanw
Other Package Authors Github handles: (comma separated, delete if none) @cyu12, @BZuckerman97, @rnluben, @abolvera
Repository:  https://github.com/rmgpanw/gtexr/tree/ropensci-software-review
Version submitted: 0.1.0.9000
Submission type: Standard

Version accepted: TBD

---

```
Package: gtexr
Title: Query the GTEx Portal API
Version: 0.1.0.9000
    person(given = "Alasdair",
           family = "Warwick",
           email = "alasdair.warwick.19@ucl.ac.uk",
    person("Benjamin", "Zuckerman", role = "aut",
    person("Abraham", "Olvera-Barrios", role = "aut",
    person("Chuin Ying", "Ung", role = "aut",
    person("Robert", "Luben", role = "aut",
    )
Description: A convenient R interface to the Genotype-Tissue Expression (GTEx)
    Portal API. For more information on the API, see
    <https://gtexportal.org/api/v2/redoc>.
    https://github.com/rmgpanw/gtexr
    cli,
    httr2 (>= 1.0.0),
    tidyr
    curl,
    httptest2,
```

`gtexr` wraps the [Genotype-Tissue Expression (GTEx) Portal API (v2)](https://gtexportal.org/api/v2/redoc), providing a simple option for programmatic data retrieval from this rich resource.

GTEx data is extensively used across diverse areas of biomedical research, from basic regulatory genomics to clinical applications. For example, researchers use GTEx to investigate tissue-specific gene expression patterns, interpret genome-wide association study findings, and infer causal relationships between gene expression and disease. While the GTEx Portal provides a web interface for data access, many analyses require programmatic access to query multiple genes, variants, or tissues systematically.

There are no other similar R packages on CRAN (and no Python packages on PyPI that I have found). I am aware of the R package [gtexRNA](https://github.com/sigven/gtexRNA/) on GitHub from [sigven](https://github.com/sigven), however it appears this was not completed, is no longer actively maintained, and was designed to query the now discontinued GTEx API V1.

Yes

> *Please note `gtexr` version 0.1.0 is already available on CRAN*

> *Please note, I do intend to submit an article about `gtexr` to the [Journal of Open-Source Software](https://joss.theoj.org/) (JOSS). The draft manuscript is located at `inst/paper` on the [ropensci-software-review](https://github.com/rmgpanw/gtexr/tree/ropensci-software-review/inst/paper) branch.*
