### 2025-02-11 #684 @mpadge
Thanks @rmgpanw, our Editor-in-Chief will respond soon, but in the meantime could you please update the "*Repository*" URL in the initial template from https://github.com/rmgpanw/gtexr to https://github.com/rmgpanw/gtexr/tree/ropensci-software-review ? Our check system will then perform checks on the appropriate branch. (Oh, and you'll need to once again call `check package` to trigger those.) Sorry for previous lack of clarity on that point; feel free to comment on [this pull request](https://github.com/ropensci/dev_guide/pull/869) which updates our submission instructions.
