### 2025-02-12 #684 @maurolepore
For the record here are my preliminary checks to understand what I would be presenting to the editorial board for consideration. If the package is taken for review then a handling editor will make their own checks and those would overwrite mine.

---

      > The "case for the package" is not made. This doesn't seem like a problem because the goal of the package is focused and well defined. But then that also means the "case for the package" seems easy to make, so it might be worth adding a sentence at the top of RADME, e.g. "The data is rich and useful, and there's an API but no R package."
      > Yes.
      > Yes.

  > Seems to not need authentication.
  > CRAN check results show a NOTE but seems easy to solve (https://cran.r-project.org/web/checks/check_results_gtexr.html)

---

#### Comments

* pkgcheck results: Passes all
* On CRAN: Yes, since 2024-09-19. Check results with a NOTE but easy to fix.

From a quick look:
* Dependencies: Few, robust.
* Size: ~ 180 R/*.R fiels but each is small.
* File names follow R/f.R and tests-f.R: Yes
* Style follows the tidyverse style guide: Yes

Attention!

* Note non-default branch. See repo url at the top comment.
* This file has no "test-" prefix: https://github.com/rmgpanw/gtexr/blob/ropensci-software-review/tests/testthat/get_significant_single_tissue_isqtls.R
* Arguments are passed implicitely (see https://github.com/rmgpanw/gtexr/blob/ropensci-software-review/R/gtex_query.R).
  ```r
  # Example: https://github.com/rmgpanw/gtexr/blob/ropensci-software-review/R/get_dataset_info.R
  get_dataset_info <- function(datasetId = "gtex_v8",
                               organizationName = "GTEx Consortium") {
    gtex_query(endpoint = "metadata/dataset")
  }
  ```
  > Make inputs explicit: A function is easier to understand if its output depends only on its inputs (i.e. its arguments). If a function returns different results with the same inputs, then some inputs must be implicit, typically because the function relies on an option or some locale setting. -- https://design.tidyverse.org/inputs-explicit.html#how-can-i-remediate-the-problem
* Uses `subset()` programmatically (see ?subset warning).
```
# ?subset
> Warning:

     This is a convenience function intended for use interactively.
     For programming it is better to use the standard subsetting
     functions like ‘[’, and in particular the non-standard evaluation
     of argument ‘subset’ can have unanticipated consequences.
```
