### 2025-02-16 #684 @adamhsparks
# `gtexr` - package editor checks

## **Editor:** [\@adamhsparks](https://github.com/adamhsparks)

### Date:
**`r cat(sprintf("**Last updated:** %s", Sys.Date()))`**

***

This report documents the editor checks of **rOpenSci** submitted package:

## Package info

**Description:**

A convenient R interface to the Genotype-Tissue Expression (GTEx)
    Portal API. For more information on the API, see
    <https://gtexportal.org/api/v2/redoc>.

**Author:** `r c(
    person(given = "Alasdair",
           family = "Warwick",
           role = c("aut", "cre", "cph"),
           email = "alasdair.warwick.19@ucl.ac.uk",
           comment = c(ORCID = "0000-0002-0800-2890")),
    person("Benjamin", "Zuckerman", role = "aut",
           comment = c(ORCID = "0000-0002-0077-6074")),
    person("Abraham", "Olvera-Barrios", role = "aut",
           comment = c(ORCID = "0000-0002-3305-4465")),
    person("Chuin Ying", "Ung", role = "aut",
           comment = c(ORCID = "0000-0001-8487-4589")),
    person("Robert", "Luben", role = "aut",
           comment = c(ORCID = "0000-0002-5088-6343"))
    )`

**repo url:** <https://github.com/rmgpanw/gtexr/tree/ropensci-software-review>

**website url:** <https://rmgpanw.github.io/gtexr/>

## Review info

#### See [editors guidelines](https://ropensci.github.io/dev_guide/editorguide.html) for further information on the rOpenSci editors checks process.

**key editor checks:**

***

### session info

```{r sessionInfo}
sessionInfo()
```

```r
R version 4.4.2 (2024-10-31)
Platform: aarch64-apple-darwin20
Running under: macOS Sequoia 15.3.1

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: Australia/Perth
tzcode source: internal

attached base packages:

 [1] miniUI_0.1.1.1    jsonlite_1.8.9    compiler_4.4.2    promises_1.3.2    Rcpp_1.0.14
 [6] colorout_1.3-2    jquerylib_0.1.4   later_1.4.1       credentials_2.0.2 yaml_2.3.10
[11] fastmap_1.2.0     mime_0.12         R6_2.6.0          curl_6.2.0        knitr_1.49
[16] htmlwidgets_1.6.4 profvis_0.4.0     openssl_2.3.2     shiny_1.10.0      bslib_0.9.0
[21] rlang_1.1.5       cachem_1.1.0      httpuv_1.6.15     xfun_0.50         sass_0.4.9
[26] fs_1.6.5          sys_3.4.3         pkgload_1.4.0     memoise_2.0.1     cli_3.6.3
[31] magrittr_2.0.3    digest_0.6.37     rstudioapi_0.17.1 xtable_1.8-4      remotes_2.5.0
[36] devtools_2.4.5    askpass_1.2.1     lifecycle_1.0.4   vctrs_0.6.5       glue_1.8.0
[41] evaluate_1.0.3    urlchecker_1.0.1  sessioninfo_1.2.3 pkgbuild_1.4.6    purrr_1.0.4
[46] rmarkdown_2.29    usethis_3.1.0     tools_4.4.2       ellipsis_0.3.2    htmltools_0.5.8.1
```

```{r pkg_dir, echo = F}
pkg_dir <- "~/Developer/GitHub/rOpenSci/Reviews/gtexr"
```

## Test installation

### test local `gtexr` install:

```{r test-local}
devtools::install(pkg_dir, dependencies = TRUE, build_vignettes = TRUE)
```

```{r github-rm}
remove.packages("gtexr")
```
#### **comments:**

Cleanly installs and un-installs from local directory.

***

### test install of `gtexr` from GitHub with:

```{r test-github}
devtools::install_github("rmgpanw/gtexr",
                         branch = "ropensci-software-review",
                         dependencies = TRUE,
                         build_vignettes = TRUE)
```

#### **comments:**

Cleanly installs and un-installs from GitHub repository branch.

***

## Check package integrity

### run checks on `gtexr` source:

```{r check-checks}
devtools::check(pkg_dir)
```
#### **comments:**

One NOTE found.

```
  Problems with news in ‘NEWS.md’:
   No news entries found.
```

***

### run tests on `gtexr` source:

```{r check-tests}
devtools::test(pkg_dir)
```
#### **comments:**

All tests pass.

***

### check `gtexr` for goodpractice:

```{r test-goodpractice}
goodpractice::gp(pkg_dir)
```

#### **comments:**

No flags from {goodpractice} aside from lines >80 chars.
This is something that I feel is up to the author and what they are comfortable with.
I try to keep mine <80 where possible, but realise sometimes this is impractical.

***

## Check package metadata files

### inspect

- #### [README](https://github.com/rmgpanw/gtexr/tree/ropensci-software-review)
- #### [DESCRIPTION](https://github.com/rmgpanw/gtexr/blob/ropensci-software-review/DESCRIPTION)

### spell check

```{r spell-check}
devtools::spell_check(pkg_dir)
```

#### **comments:**

##### README

My first thought is that if you're new to R you won't know what a Shiny App is.
Some more detail here about what it is and why new users would want to try it would be useful.

My second thought is that there's no really info in the beginning about what the API is for here.
Yes, there's a link, but I'd like as a user to just see at least a line about the API that this client services, _e.g._, https://github.com/ropensci/nasapower?tab=readme-ov-file#about-nasapower.
You can ignore the preceeding paragraph, it's there for very specific reasons that came after I wrote the one linked to here.

I would prefer to use (and am in the process of changing my READMEs over where appropriate) `pak::pak("rmgpanw/gtexr")` in place of `devtools::install_github("rmgpanw/gtexr")`, but this is not mandatory, just merely a suggestion.

##### DESCRIPTION

CRAN will likely flag your Description field for having "A convenient R interface".
I'd add more detail to it and just start with "An API interface to the Genotype-Tissue Expression (GTEx)..." and add a bit more detail about the data that are being serviced from the API.
You could refer to my {nasapower} [DESCRIPTION](https://github.com/ropensci/nasapower/blob/main/DESCRIPTION) file if you'd like.

##### Spelling

There is no Language field in the DESCRIPTION.

Several words were flagged as being misspelled.
Most don't appear to be, but there is no WORDLIST in the package that I can find to tell the checks what is spelled properly that's not in a normal dictionary.
This can be set up with `usethis::use_spell_check()`.

***

## Review test suite:

### test coverage

```{r pkg_coverage}
covr::package_coverage(pkg_dir)

```

### inspect [tests](https://github.com/rmgpanw/gtexr/tree/ropensci-software-review/tests/testthat)

#### **comments:**

Coverage is 100%, great job!

Tests appear to use mocking to avoid hitting the API from CRAN, which is also good practice.

***

### Reviewer assignment response:

rmgpanw, could you please add the rOpenSci under review badge to your README?

I'll start looking for reviewers.

---

Reviewers:
Due date:

***
