### 2025-02-16 abcd0db
delete redundant (duplicated) test file
