### 2025-02-16 85fce1e
add spellcheck and update DESCRIPTION
