+++ DESCRIPTION
@@ -17,13 +17,20 @@ Authors@R: c(
            comment = c(ORCID = "0000-0002-5088-6343"))
     )
 Description: A convenient R interface to the Genotype-Tissue Expression (GTEx) 
-    Portal API. For more information on the API, see 
+    Portal API. The GTEx project is a comprehensive public resource for studying 
+    tissue-specific gene expression and regulation in human tissues. Through 
+    systematic analysis of RNA sequencing data from 54 non-diseased tissue sites 
+    across nearly 1000 individuals, GTEx provides crucial insights into the 
+    relationship between genetic variation and gene expression. This data is 
+    accessible through the GTEx Portal API enabling programmatic access to human 
+    gene expression data. For more information on the API, see 
     <https://gtexportal.org/api/v2/redoc>.
 License: MIT + file LICENSE
 URL: https://rmgpanw.github.io/gtexr/,
     https://github.com/rmgpanw/gtexr
 BugReports: https://github.com/rmgpanw/gtexr/issues
 Encoding: UTF-8
+Language: en-US
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 Imports: 
@@ -39,6 +46,7 @@ Suggests:
     httptest2,
     knitr,
     rmarkdown,
+    spelling,
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
 VignetteBuilder: knitr
