### 2025-02-16 #684 @rmgpanw
Hi @adamhsparks many thanks to you and @maurolepore for all your insightful comments so far. I am also excited to work with you and the reviewers 👍

For now, I have:

- Updated README:

  * Included the rOpenSci Software Peer Review badge
  * Added more detail regarding GTEx
  * Installation instructions for development version amended to `pak::pak("rmgpanw/gtexr")`

- Updated DESCRIPTION with more detail regarding GTEx
- Removed duplicated test file `get_significant_single_tissue_isqtls.R`
- Set up spell checking with `usethis::use_spell_check()`
- Replaced `subset()` in `gtex_query.R`

Looking forward to making further improvements together
