### 2025-03-02 #684 @ernestguevarra
@rmgpanw and colleagues, thank you for your great work on this package. Overall, package is performant and a lot of thought has been put in the design and engineering of the functions to avoid duplication and to adhere to DRY principles. Following is my detailed review.

Let me know if you have any questions or clarifications.

Estimated hours spent reviewing: 8 hours spread over 3 days

---

### Review Info

#### session info

My review session info are as follows:

```R
Platform: x86_64-pc-linux-gnu
Running under: Pop!_OS 22.04 LTS

BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3
LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so;  LAPACK version 3.10.0

 [1] LC_CTYPE=en_GB.UTF-8       LC_NUMERIC=C               LC_TIME=en_GB.UTF-8        LC_COLLATE=en_GB.UTF-8
 [5] LC_MONETARY=en_GB.UTF-8    LC_MESSAGES=en_GB.UTF-8    LC_PAPER=en_GB.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=en_GB.UTF-8 LC_IDENTIFICATION=C

time zone: Europe/London
tzcode source: system (glibc)

[1] httptest2_1.1.0  gtexr_0.1.0.9000 testthat_3.2.3   devtools_2.4.5   usethis_2.2.3

 [1] tidyselect_1.2.1        dplyr_1.1.4             blob_1.2.4              fastmap_1.2.0           lazyeval_0.2.2
 [6] xopen_1.0.1             promises_1.3.2          rex_1.2.1               digest_0.6.37           mime_0.12
[11] lifecycle_1.0.4         waldo_0.6.1             ellipsis_0.3.2          processx_3.8.6          RSQLite_2.3.9
[16] magrittr_2.0.3          compiler_4.4.2          rlang_1.1.5             sass_0.4.9              tools_4.4.2
[21] yaml_2.3.10             data.table_1.15.4       knitr_1.49              prettyunits_1.2.0       htmlwidgets_1.6.4
[26] bit_4.0.5               pkgbuild_1.4.6          curl_6.2.1              xmlparsedata_1.0.5      xml2_1.3.6
[31] pkgload_1.4.0           miniUI_0.1.1.1          covr_3.6.4              withr_3.0.2             purrr_1.0.4
[36] desc_1.4.3              hunspell_3.0.5          goodpractice_1.0.5.9000 roxygen2_7.3.2          urlchecker_1.0.1
[41] profvis_0.3.8           whoami_1.3.0            xtable_1.8-4            debugme_1.2.0           cli_3.6.4
[46] rmarkdown_2.29          crayon_1.5.3            generics_0.1.3          remotes_2.5.0.9000      rstudioapi_0.17.1
[51] parsedate_1.3.1         cranlike_1.0.3          httr_1.4.7              commonmark_1.9.2        sessioninfo_1.2.2
[56] DBI_1.2.3               cachem_1.1.0            stringr_1.5.1           vctrs_0.6.5             jsonlite_1.9.0
[61] callr_3.7.6             rcmdcheck_1.4.0         bit64_4.0.5             cyclocomp_1.1.1         jquerylib_0.1.4
[66] tidyr_1.3.1             spelling_2.3.1          glue_1.8.0              crancache_0.0.0.9001    rematch2_2.1.2
[71] ps_1.9.0                stringi_1.8.4           later_1.4.1             tibble_3.2.1            pillar_1.10.1
[76] rappdirs_0.3.3          htmltools_0.5.8.1       brio_1.1.5              praise_1.0.0            R6_2.6.1
[81] lintr_3.1.2             httr2_1.1.0             rprojroot_2.0.4         evaluate_1.0.3          shiny_1.9.1
[86] backports_1.5.0         memoise_2.0.1           httpuv_1.6.15           bslib_0.9.0             Rcpp_1.0.14
[91] xfun_0.51               fs_1.6.5                pkgconfig_2.0.3
```

#### Test installation

Local install and GitHub install works as expected.

#### devtools::check() and devtools::test()

On `devtools::check()`, I see:

```R
── R CMD check results ──────────────────────────────────────────────────────────────────────────────────── gtexr 0.1.0.9000 ────
Duration: 37.7s

❯ checking package subdirectories ... NOTE
  Problems with news in ‘NEWS.md’:
  No news entries found.

0 errors ✔ | 0 warnings ✔ | 1 note ✖
```

On `devtools::test()`, I see:

```R
══ Results ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════

```

#### Check package metadata files

* README

    * has already been edited as per feedback by @adamhsparks. Changes made are helpful and betterintroduces the package within the context of the GTEx project and its API.
    * SUGGESTION: Might be good to have a section in the README (maybe after the introductory paragraph and before the installation section) where the different functionality groupings of the package are listed out (in bullets) with as brief/succinct a description included. This can be follow the endpoint groupings used for the reference page for the functions. This might be helpful for the potential user who goes to the GitHub repository and/or the package website to see a headline overview of what functionalities are available through the package and will complement the detailed information provided in the Getting Started vignette/page.
    * SUGGESTION: Might be good to have a final section in the README about **Community guidelines** that guides the user to the link to where issues and/or support requests can be filed, link to details on how to contribute, and link to details on the code of conduct that governs participation in the project. As an example, in my packages, I have the following:

```md

Feedback, bug reports, and feature requests are welcome; file issues or seek support [here](LINK TO GITHUB REPO ISSUES PAGE). If you would like to contribute to the package, please see our [contributing guidelines](LINK TO PACKAGE WEBSITE PAGE FOR CONTRIUBTING GUIDELINES).

This project is released with a [Contributor Code of Conduct](LINK TO THE CODE OF CONDUCT PAGE OF THE PACKAGE WEBSUTE). By contributing to this project, you agree to abide by its terms.
```

An example of how I implement this is here: https://github.com/nutriverse/sleacr#community-guidelines.

* DESCRIPTION

    * changes suggested by @adamhsparks with regard to the package title and the short description have already been incorporated by authors;
    * noted that there is no minimum R version dependency indicated in DESCRIPTION but the package uses the base pipe which was introduced in R v4.1.0; realised that packages is already on CRAN so checked the CRAN checks page (https://cloud.r-project.org/web/checks/check_results_gtexr.html) and saw this:

```R
Check: DESCRIPTION meta-information
Result: NOTE
    Missing dependency on R >= 4.1.0 because package code uses the pipe
    |> or function shorthand \(...) syntax added in R 4.1.0.
    File(s) using such syntax:
      ‘calculate_ieqtls.R’ ‘calculate_isqtls.R’ ‘download.R’
      ‘get_annotation.R’ ‘get_downloads_page_data.R’ ‘get_file_list.R’
      ‘get_gene_expression.R’ ‘get_genomic_features.R’ ‘get_image.R’
      ‘get_sample_biobank_data.R’ ‘get_service_info.R’
      ‘get_significant_single_tissue_eqtls_by_location.R’
      ‘get_single_nucleus_gex.R’ ‘get_tissue_site_detail.R’ ‘gtex_query.R’
      ‘handle_http_errors.R’ ‘utils.R’ ‘validate_args.R’
Flavors: r-devel-linux-x86_64-debian-clang, r-devel-linux-x86_64-debian-gcc, r-devel-linux-x86_64-fedora-clang, r-devel-linux-x86_64-fedora-gcc, r-devel-macos-x86_64, r-devel-windows-x86_64
```

It will be good to set a minimum R version dependency to R >= 4.1.0 to be able to address this NOTE especially if an updated version after rOpenSci review will be submitted to CRAN. I would recommend setting the minimum R version dependency to R >= 4.2.0 which is the versio in which the base pipe placeholder functionality was added. This affords the authors flexibility to use the placeholder functionality now and in the future (given that they have already been using the function shorthan approach).

* CONTRIBUTING and CODE OF CONDUCT

    * package uses/borrows liberally from the `tidyverse` text for CONTRIBUTING. Whilst the code specific steps have been edited to be specific to this package
    * SUGGESTION: edit text in the CONTRIBUTING.md file that seems to say that this pacakge is a tidyverse pacakge (first section/first paragraph) while also making sure to indicate that this package is following/emulating `tidyverse` style.
    * the contributing document has a final section for CODE OF CONDUCT with a link to a code of conduct. But when code of conduct link is tapped, it leads to a 404 page not found error within Github (https://github.com/rmgpanw/gtexr/blob/ropensci-software-review/.github/CODE_OF_CONDUCT.md). This is likely because the package itself doesn't contain/include a CODE_OF_CONDUCT.md file.
    * SUGGESTION: add a CODE_OF_CONDUCT.md file in the `.github` directory; this can be done using `usethis::use_code_of_conduct()`

#### Check documentation

* Install instructions

Install instructions are strightforward and clear. Authors have changed/shifted to using `pak::pak()` for GitHub install as suggested by @adamhsparks. Suggest to consider already adding instructions for installing via R Universe as most likely this package will get added to the rOpenSci R Universe once accepted.

* Function documentation

    * Parameters/arguments definitiion are consistent with API documentation which is helpful. Generally, the API documentation of the parameters is very clear.
    * good use of @family tag to group functions appropriately/logically and allows for these functions to be grouped in the package website. However, most of these grouped functions have the exact same parameters/arguments but they get their own separate documentation page hence there is a lot of repetition of parameters/argument documentation.
    * SUGGESTION: Might be good to consider using @rdname grouping so that these functions of the same family and share the same parameters/arguments can share the same documentation page (see https://r-pkgs.org/man.html#sec-man-multiple-functions).
    * SUGGESTION: Adding some more nuance/detail/description of the return value/output of each function might be helfpul. Currently, majority of functions just say "A tibble". This will be particularly helpful if fields of the output/returned tibble vary from each other and/or that there are specific fields that are specific to a kind of output for a particular function. If majority of functions output tibbles with the same field, to avoid repeating this in the @return tag/section, it might be good to have a vignette just describing the tibble that gets outputted across the different functions and then this vignette can be referred to via a link in the @return tag/section.
    * SUGGESTION: Might be good to use @returns instead of @return to make it consistent with @params and @examples tag/section (see https://r-pkgs.org/man.html#sec-man-returns specifically the footnote number 10).

* Vignettes

    * package has 2 vignettes; first is the main package vignette that produces the Getting Started page in the package website; This page builds on the README and provides various examples of contexts/applications that the package functions can be used. The second vignette is specific to developers and gives a bullet list of concepts and guidance that developers who want to contribute to the package need to consider and be aware of.
    * SUGGESTION: might be good to have a vignette that describes the output tibble common across multiple functions from the point of view of describing the fields that are produced/outputted. This is related to an earlier suggestion of further enriching the @returns tag/section in the function documentation. If the tibble output of functions contain consistent structure and fields, then this vignette will allow to give details on all those in one set of documentation that the function documentation can refer to without having to repeat the actual documentation.

* Examples - examples are succinct, to the point, and are clear. There is also guidance provided in vignettes for developers with regard on how to be consisent with defining examples.

#### Test functionality

* user interface improvement

An improvement/feature that the authors might consider is allowing users to specify whether to retrieve/output raw JSON from an API call. The `gtex_query()` function has a `return_raw` argument that is set to FALSE by default. In the functions that wrap `gtex_query()`, `return_raw` is set to TRUE without a user option to set to FALSE. The wrapper functions already implements certain data processing and rectangularisation of JSON output steps and always returns a `tibble`. This approach is likely very convenient for most users as a `tibble` output is what most users will be familiar dealing with. However, some users might prefer to get a raw JSON output and process this themselves for a specific use case. It will be very challenging for the package authors to try to think through all the possible use cases that package users will require and cater for each one. The processing that they currently apply to the raw JSON output are likely the most common use case. Adding an argument that gives users a choice of output extends the usability for those that may need a different processing approach to the raw JSON output.

Another improvement that may be useful for a subset of users is being able to retrieve all query results in one call without pagination. There may be some uses cases in wich users would like to be able to get all the results of a query and if the results go beyond a page, then multiple calls will need to be made to get all results. A convenience function that allows for retrieval of results from multiple pages might be useful? Doing so might need to consider API rate limits (not clear from documentation whether GTEx API has rate limits) and as such might need to throttle API requests accordingly.

* performance improvements

Functions are performant and fast. Because of pagination, size of output is fairly consistent so function calls output results quickly. If authors consider the functionality of retrieving multiple page query results, then consideration for multiple API calls performance may need to be factored in development.

#### Inspect code

Overall, the package authors have used a highly modular approach to package building and documentation. As such, package adheres quite well to DRY principles. The `gtex_query()` (unexported) is the function that drives the interface to the API and is used within all the exported functions. The GTEx API currently doesn't need any authentication so user interface is straightforward. The modular approach also helps with long-term maintenance. API changes in terms of deprecation and/or addition of parameters/arguments/queries will just require the addition of the new parameter/query to the `getxr_arguments()` function. Update to new argument documentation will also be facilitated by changes to this function. Accommodating new endpoints will also be relatively easy to implement as this will just be another function that wraps the `gtex_query()` function.

Function and variable naming has been mapped to the API endpoint names and endpoint paramters which makes a lot of sense and makes package development and maintenance much easier.

Dependency management is well implemented and consistent. Only thing to consider editing/checking is the dependency to a minimum R version as described above to address the NOTEs in CRAN checks.

#### Review test suite

Coverage is 100%! Well done. Test use mocking to avoid hitting the API when testing in CRAN.
