### 2025-03-15 #684 @zkamvar
- I have no current nor past working relationship with the author

---

The package is easy to install and straightforward in its promise to deliver an R interface to the GTEx portal. The design decision to standardize all outputs into a tibble I believe was a correct one as the results of this API appear to lend themselves well to this and it allows convenient use of tidy tools.

The app provided a really nice interface to check the performance of the examples and I really appreciated the developer's vignette that described the standardization conventions used.

A couple of points for improvement that I noticed:

For `get_service_info()`, I would suggest to pass it to `tibble::glimpse()` so that the information can be easily digested.

Messaging: Every API query reports a paging info message. This is a good thing! However, the messages were cluttering the testthat output while it was running. Moreover, I would imagine that users would be aware of what they want to query and do not want the paging message associated? My suggestion would be to include a `gtexr.verbose` option (default to `TRUE`) so that you can set it to not print the message when its set to `FALSE`. This would allow you to reduce some of the messages that pop up during the tests (which can cause some pain down the road if you have a test warning or message that you don't expect later down the road and need to figure out where it came from).

For the README, I noticed that the first query has a warning associated with it. If this is an intentional demonstration, then I would suggest to notify the reader that this warning is expected as there are more than 250 entries in the eQTL Whole Blood data set.

Along those lines, I would also suggest to provide a demonstration of using the `page` argument so that it's clear how to use it in case anyone gets a hit that has over 100000 results (the maximum results per page).

Otherwise, I do want to say that this appears to be a really good example of providing a nice R interface for an API that does not require the user to know very much about the business of handling API requests.

### Session Info

```r
─ Session info ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
 os       Ubuntu 24.04.2 LTS
 system   x86_64, linux-gnu
 ui       X11
 language en_US:en
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Los_Angeles
 pandoc   2.19.2 @ /home/zhian/.local/share/r-pandoc/2.19.2/pandoc
 quarto   1.5.56 @ /usr/local/bin/quarto

─ Packages ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
 package     * version    date (UTC) lib source
 cli           3.6.4      2025-02-13 [1] RSPM (R 4.4.0)
 curl          6.2.1      2025-02-19 [1] CRAN (R 4.4.2)
 dplyr         1.1.4      2023-11-17 [1] RSPM (R 4.4.0)
 generics      0.1.3      2022-07-05 [1] RSPM (R 4.4.0)
 glue          1.8.0      2024-09-30 [1] RSPM (R 4.4.0)
 gtexr       * 0.1.0.9000 2025-03-15 [1] local
 httr2         1.1.1      2025-03-08 [1] RSPM (R 4.4.0)
 jsonlite      1.9.1      2025-03-03 [1] RSPM (R 4.4.0)
 lifecycle     1.0.4      2023-11-07 [1] RSPM (R 4.4.0)
 magrittr      2.0.3      2022-03-30 [1] RSPM (R 4.4.0)
 pillar        1.10.1     2025-01-07 [1] RSPM (R 4.4.0)
 pkgconfig     2.0.3      2019-09-22 [1] RSPM (R 4.4.0)
 purrr         1.0.4      2025-02-05 [1] RSPM (R 4.4.0)
 R6            2.6.1      2025-02-15 [1] RSPM (R 4.4.0)
 rappdirs      0.3.3      2021-01-31 [1] RSPM (R 4.4.0)
 rlang         1.1.5      2025-01-17 [1] RSPM (R 4.4.0)
 sessioninfo   1.2.3      2025-02-05 [1] RSPM (R 4.4.0)
 tibble        3.2.1      2023-03-20 [1] RSPM (R 4.4.0)
 tidyr         1.3.1      2024-01-24 [1] RSPM (R 4.4.0)
 tidyselect    1.2.1      2024-03-11 [1] RSPM (R 4.4.0)
 utf8          1.2.4      2023-10-22 [1] RSPM (R 4.4.0)
 vctrs         0.6.5      2023-12-01 [1] RSPM (R 4.4.0)
 withr         3.0.2      2024-10-28 [1] RSPM (R 4.4.0)

 [1] /home/zhian/R/x86_64-pc-linux-gnu-library/4.4
 [2] /opt/R/4.4.2/lib/R/library
 * ── Packages attached to the search path.

──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```
