### 2025-03-19 0d29995
add gtexr.verbose option; tests failing currently for functions calling paging_info_messages() outside of gtex_query()
