+++ tests/testthat/setup.R
@@ -1,1 +1,2 @@
 library(httptest2)
+options(list(gtexr.verbose = FALSE))
