### 2025-03-28 748afb3
implement `.return_raw` argument - if `TRUE`, returns raw json response
