+++ R/get_linkage_disequilibrium_data.R
@@ -20,8 +20,7 @@
 get_linkage_disequilibrium_data <- function(gencodeId,
                                             datasetId = "gtex_v8",
                                             page = 0,
-                                            itemsPerPage = 250) {
-  resp_body <- gtex_query(endpoint = "dataset/ld", return_raw = TRUE)
-
-  process_resp_body_linkage_disequilibrium(resp_body)
+                                            itemsPerPage = 250,
+                                            .return_raw = FALSE) {
+  gtex_query(endpoint = "dataset/ld", process_linkage_disequilibrium_resp_json)
 }
