+++ R/get_exons.R
@@ -24,6 +24,7 @@ get_exons <- function(gencodeId,
                       genomeBuild = NULL,
                       datasetId = "gtex_v8",
                       page = 0,
-                      itemsPerPage = 250) {
+                      itemsPerPage = 250,
+                      .return_raw = FALSE) {
   gtex_query(endpoint = "reference/exon")
 }
