+++ R/get_service_info.R
@@ -5,6 +5,8 @@
 #' [GTEx Portal API
 #' documentation](https://gtexportal.org/api/v2/redoc#tag/GTEx-Portal-API-Info/operation/get_service_info_api_v2__get).
 #'
+#' @inheritParams gtexr_arguments
+#'
 #' @return A tibble.
 #' @export
 #' @family GTEx Portal API Info
@@ -13,8 +15,12 @@
 #' \dontrun{
 #' get_service_info()
 #' }
-get_service_info <- function() {
-  gtex_query(endpoint = NULL, return_raw = TRUE) |>
+get_service_info <- function(.return_raw = FALSE) {
+  gtex_query(endpoint = NULL, process_get_service_info_resp_json)
+}
+
+process_get_service_info_resp_json <- function(resp_json) {
+  resp_json |>
     purrr::map_at("organization", tibble::as_tibble) |>
     tibble::as_tibble() |>
     tidyr::unnest(cols = "organization", names_sep = "_")
