+++ R/get_variant.R
@@ -38,6 +38,7 @@ get_variant <- function(snpId = NULL,
                         pos = NULL,
                         datasetId = "gtex_v8",
                         page = 0,
-                        itemsPerPage = 250) {
+                        itemsPerPage = 250,
+                        .return_raw = FALSE) {
   gtex_query(endpoint = "dataset/variant")
 }
