### 2025-03-29 61adeac
add `.verbose` option as an argument for relevant functions
