+++ R/get_significant_single_tissue_isqtls.R
@@ -29,6 +29,7 @@ get_significant_single_tissue_isqtls <- function(gencodeIds,
                                                  datasetId = "gtex_v8",
                                                  page = 0,
                                                  itemsPerPage = 250,
+                                                 .verbose = getOption("gtexr.verbose"),
                                                  .return_raw = FALSE) {
   gtex_query(endpoint = "association/singleTissueISqtl")
 }
