+++ R/get_collapsed_gene_model_exon.R
@@ -24,6 +24,7 @@ get_collapsed_gene_model_exon <- function(gencodeId,
                                           datasetId = "gtex_v8",
                                           page = 0,
                                           itemsPerPage = 250,
+                                          .verbose = getOption("gtexr.verbose"),
                                           .return_raw = FALSE) {
   gtex_query(endpoint = "dataset/collapsedGeneModelExon")
 }
