+++ R/get_sqtl_genes.R
@@ -24,6 +24,7 @@ get_sqtl_genes <- function(tissueSiteDetailId,
                            datasetId = "gtex_v8",
                            page = 0,
                            itemsPerPage = 250,
+                           .verbose = getOption("gtexr.verbose"),
                            .return_raw = FALSE) {
   gtex_query(endpoint = "association/sgene")
 }
