+++ R/get_significant_single_tissue_eqtls.R
@@ -51,6 +51,7 @@ get_significant_single_tissue_eqtls <- function(gencodeIds = NULL,
                                                 datasetId = "gtex_v8",
                                                 page = 0,
                                                 itemsPerPage = 250,
+                                                .verbose = getOption("gtexr.verbose"),
                                                 .return_raw = FALSE) {
   gtex_query(endpoint = "association/singleTissueEqtl")
 }
