+++ DESCRIPTION
@@ -33,6 +33,7 @@ Encoding: UTF-8
 Language: en-US
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
+Depends: R (>= 4.2.0)
 Imports: 
     cli,
     dplyr,
