+++ tests/testthat/test-gtex_query.R
@@ -3,8 +3,30 @@ with_mock_dir("gtex_query", {
     expect_equal(nrow(get_service_info()), 1L)
   })
 
+  test_that("Call gtex_query() with `.return_raw = TRUE`", {
+    expect_equal(
+      get_service_info(.return_raw = TRUE),
+      list(
+        id = "org.gtexportal.rest.v2",
+        name = "GTEx Portal V2 API",
+        version = "2.0.0",
+        organization = list(name = "GTEx Project", url = "https://gtexportal.org"),
+        description = "This service provides access to the data powering the GTEx portal.",
+        contactUrl = "https://gtexportal.org/home/contact",
+        documentationUrl = "https://gtexportal.org/api/v2/docs",
+        environment = "prod"
+      )
+    )
+  })
+
   test_that("Call gtex_query() with arguments", {
     expect_equal(nrow(suppressWarnings(get_gene_search("CRP", itemsPerPage = 1))), 1L)
+  })
+
+  test_that("gtex_query() prints paging info message when `.verbose = TRUE`", {
+    expect_message(suppressWarnings(get_gene_search(
+      "CRP", itemsPerPage = 1, .verbose = TRUE
+    )))
   })
 
   test_that("gtex_query() raises a warning if total n items exceeds page size", {
@@ -23,9 +45,27 @@ test_that("`process_na_and_zero_char_query_params()` works", {
   )
 })
 
-test_that("`validate_verbose_arg()` works", {
+test_that("`validate_verbose_arg` works", {
   withr::with_options(
     list(gtexr.verbose = c("invalid", "option")),
     expect_error(get_genes("CRP"), regexp = "You supplied: character of length 2")
   )
+})
+
+test_that("`.return_raw` argument works", {
+  # check `validate_return_raw_arg()`
+  gtex_query_wrapper_fn_with_no_return_raw_arg <- function() {
+    gtex_query()
+  }
+
+  gtex_query_wrapper_fn_with_invalid_return_raw_arg <- function(.return_raw = "TRUE") {
+    gtex_query()
+  }
+
+  expect_error(gtex_query_wrapper_fn_with_no_return_raw_arg(),
+               regexp = "Missing argument: `.return_raw`")
+
+  expect_error(gtex_query_wrapper_fn_with_invalid_return_raw_arg(),
+               regexp = "You supplied: character of length 1")
+
 })
