### 2025-03-29 #684 @rmgpanw
Thank you @ernestguevarra and @zkamvar from me too, I very much appreciate the care and effort you have taken. Apologies, I am running slightly behind and will aim to post a full response in the next few days. I hope this is ok @adamhsparks.

For now, I have implemented the following new features ([ropensci-software-review](https://github.com/rmgpanw/gtexr/tree/ropensci-software-review) branch) based on your feedback:

- All functions include a `.return_raw` argument, allowing the user to retrieve the raw JSON from an API call (@ernestguevarra)
- Functions that display a 'Paging info' message now include a `.verbose` argument, providing an option to suppress these messages, This behaviour can be set globally with `options(list("gtexr.verbose" = FALSE))`
