### 2025-03-29 66ebc33
amend CONTRIBUTING.md - now states gtexr emulates tidyverse style
