+++ .github/CONTRIBUTING.md
@@ -1,7 +1,7 @@
 # Contributing to gtexr
 
 This outlines how to propose a change to gtexr.
-For a detailed discussion on contributing to this and other tidyverse packages, please see the [development contributing guide](https://rstd.io/tidy-contrib) and our [code review principles](https://code-review.tidyverse.org/).
+For a detailed discussion on contributing to this and other packages emulating tidyverse style, please see the tidyverse [development contributing guide](https://rstd.io/tidy-contrib) and [code review principles](https://code-review.tidyverse.org/).
 
 ## Fixing typos
 
