### 2025-03-29 c097596
add CODE_OF_CONDUCT.md
