+++ README.Rmd
@@ -65,3 +65,9 @@ get_significant_single_tissue_eqtls(gencodeId = c(
   "ENSG00000203782.5"
 ))
 ```
+
+## Community guidelines
+
+Feedback, bug reports, and feature requests are welcome; file issues or seek support [here](https://rmgpanw.github.io/gtexr/issues). If you would like to contribute to the package, please see our [contributing guidelines](https://rmgpanw.github.io/gtexr/CONTRIBUTING.html).
+  
+Please note that the gtexr project is released with a [Contributor Code of Conduct](https://rmgpanw.github.io/gtexr/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
