+++ R/get_dataset_info.R
@@ -3,7 +3,7 @@
 #' @description [GTEx Portal API
 #'   documentation](https://gtexportal.org/api/v2/redoc#tag/Metadata-Endpoints/operation/get_dataset_info_api_v2_metadata_dataset_get)
 #'
-#' @details Note: this current only returns empty values.
+#' @details Note: this currently only returns empty values.
 #'
 #' @inheritParams gtexr_arguments
 #'
