### 2025-03-30 16dc9e0
update documentation with overview of available functionality
