### 2025-03-30 ab512d6
update roxygen documentation to use `@returns` instead of `@return`
