+++ R/get_multi_tissue_eqtls.R
@@ -15,7 +15,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @return A tibble.
+#' @returns A tibble.
 #' @export
 #' @family Static Association Endpoints
 #'
