+++ R/get_gwas_catalog_by_location.R
@@ -8,7 +8,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @return A tibble.
+#' @returns A tibble.
 #' @export
 #' @family Reference Genome Endpoints
 #'
