+++ R/get_significant_single_tissue_eqtls.R
@@ -19,7 +19,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @return A tibble.
+#' @returns A tibble.
 #' @export
 #' @family Static Association Endpoints
 #'
