+++ R/get_gene_search.R
@@ -10,7 +10,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @return A tibble.
+#' @returns A tibble.
 #' @export
 #' @family Reference Genome Endpoints
 #'
