+++ R/calculate_splicing_quantitative_trait_loci.R
@@ -5,7 +5,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @return A tibble.
+#' @returns A tibble.
 #' @export
 #' @family Dynamic Association Endpoints
 #'
