+++ R/get_image.R
@@ -5,7 +5,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @return A tibble.
+#' @returns A tibble.
 #' @export
 #' @family Histology Endpoints
 #'
