+++ R/get_single_nucleus_gex.R
@@ -7,7 +7,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @return A tibble
+#' @returns A tibble
 #' @export
 #' @family Expression Data Endpoints
 #'
