+++ R/get_significant_single_tissue_isqtls.R
@@ -12,7 +12,7 @@
 #' [GTEx Portal API documentation](https://gtexportal.org/api/v2/redoc#tag/Static-Association-Endpoints/operation/get_significant_single_tissue_isqtls_api_v2_association_singleTissueISqtl_get).
 #'
 #' @inheritParams gtexr_arguments
-#' @return A tibble
+#' @returns A tibble
 #' @export
 #' @family Static Association Endpoints
 #'
