### 2025-03-30 4ba426d
enable `itemsPerPage` arg to be set globally using option "gtexr.itemsPerPage"
