+++ R/get_top_expressed_genes.R
@@ -23,7 +23,7 @@ get_top_expressed_genes <- function(tissueSiteDetailId,
                                     datasetId = "gtex_v8",
                                     filterMtGene = TRUE,
                                     page = 0,
-                                    itemsPerPage = 250,
+                                    itemsPerPage = getOption("gtexr.itemsPerPage"),
                                     .verbose = getOption("gtexr.verbose"),
                                     .return_raw = FALSE) {
   gtex_query(endpoint = "expression/topExpressedGene")
