+++ R/get_subject.R
@@ -23,7 +23,7 @@ get_subject <- function(datasetId = "gtex_v8",
                         hardyScales = NULL,
                         subjectIds = NULL,
                         page = 0,
-                        itemsPerPage = 250,
+                        itemsPerPage = getOption("gtexr.itemsPerPage"),
                         .verbose = getOption("gtexr.verbose"),
                         .return_raw = FALSE) {
   gtex_query(endpoint = "dataset/subject")
