+++ R/get_full_get_collapsed_gene_model_exon.R
@@ -18,7 +18,7 @@
 #' }
 get_full_get_collapsed_gene_model_exon <- function(gencodeId,
                                                    page = 0,
-                                                   itemsPerPage = 250,
+                                                   itemsPerPage = getOption("gtexr.itemsPerPage"),
                                                    .verbose = getOption("gtexr.verbose"),
                                                    .return_raw = FALSE) {
   gtex_query(endpoint = "dataset/fullCollapsedGeneModelExon")
