+++ R/get_gene_search.R
@@ -22,7 +22,7 @@ get_gene_search <- function(geneId,
                             gencodeVersion = "v26",
                             genomeBuild = "GRCh38/hg38",
                             page = 0,
-                            itemsPerPage = 250,
+                            itemsPerPage = getOption("gtexr.itemsPerPage"),
                             .verbose = getOption("gtexr.verbose"),
                             .return_raw = FALSE) {
   gtex_query(endpoint = "reference/geneSearch")
