+++ R/get_significant_single_tissue_isqtls.R
@@ -28,7 +28,7 @@ get_significant_single_tissue_isqtls <- function(gencodeIds,
                                                  tissueSiteDetailIds = NULL,
                                                  datasetId = "gtex_v8",
                                                  page = 0,
-                                                 itemsPerPage = 250,
+                                                 itemsPerPage = getOption("gtexr.itemsPerPage"),
                                                  .verbose = getOption("gtexr.verbose"),
                                                  .return_raw = FALSE) {
   gtex_query(endpoint = "association/singleTissueISqtl")
