+++ R/get_multi_tissue_eqtls.R
@@ -31,7 +31,7 @@ get_multi_tissue_eqtls <- function(gencodeIds,
                                    variantId = NULL,
                                    datasetId = "gtex_v8",
                                    page = 0,
-                                   itemsPerPage = 250,
+                                   itemsPerPage = getOption("gtexr.itemsPerPage"),
                                    .verbose = getOption("gtexr.verbose"),
                                    .return_raw = FALSE) {
   gtex_query(endpoint = "association/metasoft")
