+++ R/get_eqtl_genes.R
@@ -32,7 +32,7 @@
 get_eqtl_genes <- function(tissueSiteDetailIds,
                            datasetId = "gtex_v8",
                            page = 0,
-                           itemsPerPage = 250,
+                           itemsPerPage = getOption("gtexr.itemsPerPage"),
                            .verbose = getOption("gtexr.verbose"),
                            .return_raw = FALSE) {
   gtex_query(endpoint = "association/egene")
