+++ R/get_transcripts.R
@@ -23,7 +23,7 @@ get_transcripts <- function(gencodeId,
                             gencodeVersion = "v26",
                             genomeBuild = "GRCh38/hg38",
                             page = 0,
-                            itemsPerPage = 250,
+                            itemsPerPage = getOption("gtexr.itemsPerPage"),
                             .verbose = getOption("gtexr.verbose"),
                             .return_raw = FALSE) {
   gtex_query(endpoint = "reference/transcript")
