+++ R/get_news_item.R
@@ -16,7 +16,7 @@
 #' get_news_item()
 #' }
 get_news_item <- function(page = 0,
-                          itemsPerPage = 250,
+                          itemsPerPage = getOption("gtexr.itemsPerPage"),
                           .verbose = getOption("gtexr.verbose"),
                           .return_raw = FALSE) {
   gtex_query(endpoint = "admin/newsItem")
