### 2025-03-30 c7e43bc
improve warning message and README documentation regarding itemsPerPage arg
