### 2025-03-31 4f3ef48
update README with instructions for installing dev version from R Universe
