### 2025-03-31 51d94dd
update README with 'Citing gtexr' section
