+++ README.Rmd
@@ -87,6 +87,14 @@ get_significant_single_tissue_eqtls(gencodeId = c(
 ))
 ```
 
+## Citing gtexr
+
+If you find gtexr useful, please consider citing both GTEx and gtexr. Citation details are available [here](https://rmgpanw.github.io/gtexr/authors.html#citation).
+
+Example publications citing gtexr include:
+
+- [Schwartz R, Warwick AN, et al. Genetic Distinctions Between Reticular Pseudodrusen and Drusen: A Genome-Wide Association Study. Am J Ophthalmol. 2025 Mar 8:S0002-9394(25)00119-9. doi: 10.1016/j.ajo.2025.03.007. Epub ahead of print. PMID: 40064387](https://www.ajo.com/article/S0002-9394(25)00119-9/fulltext)
+
 ## Community guidelines
 
 Feedback, bug reports, and feature requests are welcome; file issues or seek support [here](https://rmgpanw.github.io/gtexr/issues). If you would like to contribute to the package, please see our [contributing guidelines](https://rmgpanw.github.io/gtexr/CONTRIBUTING.html).
