+++ R/get_eqtl_genes.R
@@ -29,7 +29,7 @@
 #' \dontrun{
 #' get_eqtl_genes(c("Whole_Blood", "Artery_Aorta"))
 #' }
-get_eqtl_genes <- function(tissueSiteDetailIds,
+get_eqtl_genes <- function(tissueSiteDetailIds = NULL,
                            datasetId = "gtex_v8",
                            page = 0,
                            itemsPerPage = getOption("gtexr.itemsPerPage"),
