+++ man/get_sqtl_genes.Rd
@@ -5,7 +5,7 @@
 \title{Get Sqtl Genes}
 \usage{
 get_sqtl_genes(
-  tissueSiteDetailId,
+  tissueSiteDetailIds = NULL,
   datasetId = "gtex_v8",
   page = 0,
   itemsPerPage = getOption("gtexr.itemsPerPage"),
@@ -14,9 +14,9 @@ get_sqtl_genes(
 )
 }
 \arguments{
-\item{tissueSiteDetailId}{String. The ID of the tissue of interest. Can be a
-GTEx specific ID (e.g. "Whole_Blood"; use \code{\link[=get_tissue_site_detail]{get_tissue_site_detail()}}
-to see valid values) or an Ontology ID.}
+\item{tissueSiteDetailIds}{Character vector of IDs for tissues of interest.
+Can be GTEx specific IDs (e.g. "Whole_Blood"; use
+\code{\link[=get_tissue_site_detail]{get_tissue_site_detail()}} to see valid values) or Ontology IDs.}
 
 \item{datasetId}{String. Unique identifier of a dataset. Usually includes a
 data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".}
