### 2025-03-31 7b06482
fix default argument values for get_sample_datasets_endpoints() to match API defaults
