+++ R/get_sample_datasets_endpoints.R
@@ -33,10 +33,10 @@ get_sample_datasets_endpoints <- function(datasetId = "gtex_v8",
                                           rin = NULL,
                                           uberonIds = NULL,
                                           dataTypes = NULL,
-                                          sortBy = NULL,
-                                          sortDirection = NULL,
-                                          page = NULL,
-                                          itemsPerPage = NULL,
+                                          sortBy = "sampleId",
+                                          sortDirection = "asc",
+                                          page = 0,
+                                          itemsPerPage = getOption("gtexr.itemsPerPage"),
                                           .verbose = getOption("gtexr.verbose"),
                                           .return_raw = FALSE) {
   gtex_query(endpoint = "dataset/sample")
