### 2025-03-31 924f8fe
update developed_guide.Rmd vignette to detail `.return_raw`, `.verbose` and `itemsPerPage` arguments
