### 2025-03-31 a77721b
update NEWS.md
