### 2025-03-31 410a785
update @returns to detail that a list is returned if `.return_raw = TRUE`
