+++ R/get_variant.R
@@ -16,7 +16,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @returns A tibble.
+#' @returns A tibble. Or a list if `.return_raw = TRUE`.
 #' @export
 #' @family Datasets Endpoints
 #'
