+++ R/get_sample_biobank_data.R
@@ -5,7 +5,7 @@
 #' documentation](https://gtexportal.org/api/v2/redoc#tag/Biobank-Data-Endpoints/operation/get_sample_api_v2_biobank_sample_get)
 #'
 #' @inheritParams gtexr_arguments
-#' @returns A tibble
+#' @returns A tibble. Or a list if `.return_raw = TRUE`.
 #' @export
 #' @family Biobank Data Endpoints
 #'
