+++ R/get_median_junction_expression.R
@@ -12,7 +12,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @returns A tibble.
+#' @returns A tibble. Or a list if `.return_raw = TRUE`.
 #' @export
 #' @family Expression Data Endpoints
 #'
