+++ R/get_linkage_disequilibrium_data.R
@@ -11,7 +11,7 @@
 #' documentation](https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_variant_by_location_api_v2_dataset_variantByLocation_get)
 #'
 #' @inheritParams gtexr_arguments
-#' @returns A tibble.
+#' @returns A tibble. Or a list if `.return_raw = TRUE`.
 #' @export
 #' @family Datasets Endpoints
 #'
