+++ R/get_median_transcript_expression.R
@@ -13,7 +13,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @returns A tibble.
+#' @returns A tibble. Or a list if `.return_raw = TRUE`.
 #' @export
 #' @family Expression Data Endpoints
 #'
