+++ R/get_significant_single_tissue_eqtls.R
@@ -19,7 +19,7 @@
 #'
 #' @inheritParams gtexr_arguments
 #'
-#' @returns A tibble.
+#' @returns A tibble. Or a list if `.return_raw = TRUE`.
 #' @export
 #' @family Static Association Endpoints
 #'
