### 2025-03-31 #684 @rmgpanw
Thank you @ernestguevarra and @zkamvar for your insightful comments, which will certainly improve the quality of this package. Please find my responses below - I look forward to your thoughts.
## ernestguevarra

-  [x] Allow users to specify whether to retrieve raw JSON from an API call, by setting argument `.return_raw` to `FALSE`. Commit [748afb3](https://github.com/rmgpanw/gtexr/commit/748afb3b7f328b4177d1426b44faa8a85ec5f027)
-  [x] Update DESCRIPTION - specify minimum R version dependency >= 4.2.0. Commit [2af9b8b](2af9b8b2423f3c7f572de045bc056d94e6b75452)
-  [x] Functions now use the roxygen tag `@returns`, instead of `@return`. Commit [ab512d6](ab512d61adb64ed2dbb005d88cb597024deb3c1a)
- Documentation
	-  [x] A summary of available functionality describing the different API endpoint groupings, has been added to the README and pkgdown site reference page. Commit [16dc9e0](16dc9e03e0bf04be1dfb45708f640401a6eaf416)
	-  [x] Community guideline section now added to the README, including a link to `CODE_OF_CONDUCT.md`. Commit [c097596](c097596ea272cb918365cb621a17602097167166)
	-  [x] README updated with installation instructions from R Universe. Commit [4f3ef48](4f3ef486be004dcd1e1cba0582a71d35b70bc31e)
	-  [x] `CONTRIBUTING.md` amended to say that gtexr *emulates* tidyverse style. Commit [66ebc33](66ebc33dfaa8ef6c2e0639644b72e92c3169cafb)
	-  [x] Documentation regarding functions that return paginated responses has been clarified. Please also see response to @zkamvar
- Other suggestions
	-  Grouping multiple related functions with `@Rdname` - while I see how this could be beneficial, the documentation would then no longer mirror the GTEx API. Overall, I would therefore prefer to hold off implementing this.
	-  Adding some more nuance/detail/description of the return value/output of each function - for now, I would prefer to stay with a simpler `@returns` description, but have updated this to indicate that a list is returned if `return_raw = TRUE` (commit [410a785](410a785f45042a6274d92947760831f2d6dd2aa1)). Ideally I would like to revisit this however as I (and hopefully other users) gain more familiarity with the complete set of GTEx API endpoints, summarising these learning points in a dedicated vignette, as you suggest.

## Zhian

-  [x] Implemented a `.verbose`, allowing users to silence paging information messages. This behaviour can be set globally with `options(list("gtexr.verbose" = FALSE))`
	-  [x] Added explanatory text for the warning message in README example. Commit [c7e43bc](c7e43bcb371b1b0c938f88ece051af41e4992d21)
	-  [x] Updated 'getting started' vignette with details for how to handle paginated responses that exceed the default items per page setting, either by adjusting `itemsPerPage` (recommended approach) or iterating through pages. Commit [95f6189](95f6189ac912663101267397df7af1c75f1e511b)
	-  [x] Updated README example call to `get_service_info()` to use `tibble::glimpse()`. Commit [c7e43bc](c7e43bcb371b1b0c938f88ece051af41e4992d21)

## Other

-  [x] Updated `NEWS.md`. Commit [a77721b](a77721b84e56c45790a8f8438ae7e3428fdbf2e2)
-  [x] Added 'citing gtexr' section to README. Commit [51d94dd](51d94dd09424bf4fab518b38523cc7d1bf586cfb)
-  [x] Update 'getting started' vignette with instructions for how to run the gtexr app locally. Commit [09d00d2](09d00d2f45978fb09a1cac9221730b4cef3110e1)
-  [x] Developer guide vignette updated to include new `.return_raw` and `.verbose` arguments, and `gtexr.itemsPerPage` option. Commit [924f8fe](924f8fea2195eea451fa4de727b538dfe1978104)
