### 2025-04-01 6d6f29f
rename `get_sample_datasets_endpoints()` to `get_sample_datasets()`
