+++ tests/testthat/test-get_sample_datasets.R
@@ -1,6 +1,6 @@
-test_that("`get_sample_datasets_endpoints()` returns tibble with expected colnames", {
+test_that("`get_sample_datasets()` returns tibble with expected colnames", {
   skip_if_offline()
-  result <- get_sample_datasets_endpoints(itemsPerPage = 1) |>
+  result <- get_sample_datasets(itemsPerPage = 1) |>
     suppressWarnings()
 
   expect_s3_class(result, "tbl_df")
