+++ man/get_linkage_disequilibrium_data.Rd
@@ -56,7 +56,7 @@ Other Datasets Endpoints:
 \code{\link{get_full_get_collapsed_gene_model_exon}()},
 \code{\link{get_functional_annotation}()},
 \code{\link{get_linkage_disequilibrium_by_variant_data}()},
-\code{\link{get_sample_datasets_endpoints}()},
+\code{\link{get_sample_datasets}()},
 \code{\link{get_subject}()},
 \code{\link{get_tissue_site_detail}()},
 \code{\link{get_variant}()},
