+++ NEWS.md
@@ -10,15 +10,20 @@ which can be set to `FALSE` to suppress pagination messages. The `itemsPerPage`
 argument for these functions can also be set globally by adjusting option
 "gtexr.itemsPerPage".
 
+* **Breaking change:** `get_sample_datasets_endpoints()` has been renamed to
+`get_sample_datasets()`. This is to match the naming convention used for 
+`get_sample_biobank_data()`, whereby 'get_sample' is appended with 
+their respective category titles 'datasets' and 'biobank_data'.
+
 ## Minor changes and bug fixes
 
 * Various function arguments have been updated to match the GTEx API:
 
   - `get_sqtl_genes()` argument `tissueSiteDetailId` has been pluralised to 
-  `tissueSiteDetailIds`
+  `tissueSiteDetailIds`.
   
-  - `get_eqtl_genes()`, `get_sqtl_genes()` and `get_sample_datasets_endpoints()` 
-  default argument values now match API
+  - `get_eqtl_genes()`, `get_sqtl_genes()` and `get_sample_datasets()` (formerly
+  called `get_sample_datasets_endpoints()`) default argument values now match API.
 
 # gtexr 0.1.0
 
