### 2025-04-02 5c131ef
add test-gtexr_api_match_checks.R and make various fixes to bring gtexr in line with GTEx API spec (summarised in NEWS.md)
