+++ R/get_sample_datasets.R
@@ -24,13 +24,13 @@ get_sample_datasets <- function(datasetId = "gtex_v8",
                                           ageBrackets = NULL,
                                           sex = NULL,
                                           pathCategory = NULL,
-                                          tissueSiteDetailId = NULL,
+                                          tissueSiteDetailIds = NULL,
                                           aliquotIds = NULL,
                                           autolysisScores = NULL,
                                           hardyScales = NULL,
-                                          ischemicTime = NULL,
+                                          ischemicTimes = NULL,
                                           ischemicTimeGroups = NULL,
-                                          rin = NULL,
+                                          rins = NULL,
                                           uberonIds = NULL,
                                           dataTypes = NULL,
                                           sortBy = "sampleId",
