+++ R/get_variant_by_location.R
@@ -23,6 +23,7 @@ get_variant_by_location <- function(start,
                                     chromosome,
                                     sortBy = "pos",
                                     sortDirection = "asc",
+                                    datasetId = "gtex_v8",
                                     page = 0,
                                     itemsPerPage = getOption("gtexr.itemsPerPage"),
                                     .verbose = getOption("gtexr.verbose"),
