+++ R/get_subject.R
@@ -6,7 +6,7 @@
 #' release.
 #'
 #' [GTEx Portal API
-#' documentation](https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_tissue_site_detail_api_v2_dataset_tissueSiteDetail_get)
+#' documentation](https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_subject_api_v2_dataset_subject_get)
 #'
 #' @inheritParams gtexr_arguments
 #' @returns A tibble. Or a list if `.return_raw = TRUE`.
@@ -20,7 +20,7 @@
 get_subject <- function(datasetId = "gtex_v8",
                         sex = NULL,
                         ageBrackets = NULL,
-                        hardyScales = NULL,
+                        hardyScale = NULL,
                         subjectIds = NULL,
                         page = 0,
                         itemsPerPage = getOption("gtexr.itemsPerPage"),
