+++ tests/testthat/test-gtexr_api_match_checks.R
@@ -1,0 +1,240 @@
+
+get_gtex_openapi_json <- function() {
+  httr2::request("https://gtexportal.org/api/v2/openapi.json") |>
+    httr2::req_user_agent("gtexr (https://github.com/rmgpanw/gtexr)") |>
+    httr2::req_perform() |>
+    httr2::resp_body_json()
+}
+
+get_gtex_api_endpoints <- function(gtex_openapi_json) {
+  result <- gtex_openapi_json$paths |>
+    purrr::map(\(x) x$get) |>
+    purrr::map(\(x) tibble::tibble(title = x$summary,
+                                   category = x$tags[[1]],
+                                   fn = x$summary |>
+                                     tolower() |>
+                                     stringr::str_replace_all(" ", "_"),
+                                   params_raw = list(x$parameters))) |>
+    dplyr::bind_rows(.id = "endpoint")
+
+  # endpoints with non-unique names - join with category title
+  duplicate_endpoint_names <- result |>
+    dplyr::count(.data[["fn"]]) |>
+    dplyr::arrange(dplyr::desc(.data[["n"]])) |>
+    dplyr::filter(.data[["n"]] > 1) |>
+    dplyr::pull(dplyr::all_of("fn"))
+
+  result <- result |>
+    dplyr::mutate(
+      "fn" = dplyr::case_when(
+        .data[["fn"]] %in% !!duplicate_endpoint_names ~ stringr::str_glue(
+          "{fn}_{stringr::str_remove(tolower(stringr::str_replace_all(category, ' ', '_')), '_endpoints')}"
+        ),
+        TRUE ~ .data[["fn"]]
+      )
+    )
+
+  # indicate whether returns paginated response
+  result <- result |>
+    dplyr::mutate(returns_paginated_response = purrr::map_lgl(params_raw, \(x) "itemsPerPage" %in% x$name))
+
+  result <- result |>
+    dplyr::mutate(params = purrr::map(
+      .data[["params_raw"]],
+      \(params_raw) purrr::map_df(
+        params_raw,
+        \(x) {
+          result <- list(
+            arg_raw = x$name,
+            value = ifelse(x$required, list(NA), list(x$schema$default)),
+            type = ifelse(is.null(x$schema$type), NA, x$schema$type),
+            required = x$required
+          )
+
+          result$arg <- dplyr::case_when(
+            result[["type"]] == "array" &
+              result[["arg_raw"]] != "pathCategory" ~ paste0(result[["arg_raw"]], "s"),
+            result[["arg_raw"]] == "featureId" ~ ".featureId",
+            TRUE ~ result[["arg_raw"]]
+          )
+
+          result
+        }
+      )
+    )) |>
+    dplyr::mutate(param_names = purrr::map(params, \(x) suppressWarnings(x$arg)))
+
+  return(result)
+}
+
+get_gtexr_fn_args <- function() {
+  tibble::tibble(fn = getNamespaceExports("gtexr")) |>
+    dplyr::mutate(args_all = purrr::map(
+      .data[["fn"]],
+      \(fn) fn |>
+        get(envir = asNamespace("gtexr"), inherits = FALSE) |>
+        rlang::fn_fmls() |>
+        as.list() |>
+        purrr::map_if(.p = rlang::is_missing, \(x) NA) |>
+        purrr::map_at("itemsPerPage", \(x) 250) |>
+        tibble::enframe(name = "arg", value = "value")
+    )) |>
+    dplyr::mutate("has_verbose_arg" = purrr::map_lgl(.data[["args_all"]], \(args) ".verbose" %in% args$arg)) |>
+    dplyr::mutate("has_return_raw_arg" = purrr::map_lgl(.data[["args_all"]], \(args) ".return_raw" %in% args$arg)) |>
[…]
