### 2025-04-02 833ceac
update R-CMD-check.yaml to automatically run once a month
