+++ .github/workflows/R-CMD-check.yaml
@@ -5,6 +5,8 @@ on:
     branches: [main, master]
   pull_request:
     branches: [main, master]
+  schedule:
+    - cron: '0 0 1 * *'  # Run at 00:00 on the first day of every month
 
 name: R-CMD-check
 
