### 2025-04-03 fd2cd82
trim redundant code from gtex_query()
