+++ R/get_multi_tissue_eqtls.R
@@ -23,6 +23,12 @@
 #' \dontrun{
 #' # search by gene
 #' get_multi_tissue_eqtls(gencodeId = "ENSG00000132693.12")
+#'
+#' # note that 'tissues' is a list-column
+#' x <- get_multi_tissue_eqtls(gencodeId = "ENSG00000132693.12",
+#'                             variantId = "chr1_159476920_T_C_b38")
+#'
+#' x$tissues[[1]]
 #' }
 get_multi_tissue_eqtls <- function(gencodeId,
                                    variantId = NULL,
@@ -31,5 +37,16 @@ get_multi_tissue_eqtls <- function(gencodeId,
                                    itemsPerPage = getOption("gtexr.itemsPerPage"),
                                    .verbose = getOption("gtexr.verbose"),
                                    .return_raw = FALSE) {
-  gtex_query(endpoint = "association/metasoft")
+  gtex_query(endpoint = "association/metasoft", process_get_multi_tissue_eqtls_resp_json)
+}
+
+process_get_multi_tissue_eqtls_resp_json <- function(resp_json) {
+  resp_json$data |>
+    purrr::map(\(x) x |>
+                 purrr::map_at("tissues", \(tissue) list(
+                   purrr::map(tissue, tibble::as_tibble) |>
+                     dplyr::bind_rows(.id = "tissue")
+                 )) |>
+                 tibble::as_tibble()) |>
+    dplyr::bind_rows()
 }
