+++ man/get_multi_tissue_eqtls.Rd
@@ -56,6 +56,12 @@ By default, the service queries the latest GTEx release. The retrieved data is s
 \dontrun{
 # search by gene
 get_multi_tissue_eqtls(gencodeId = "ENSG00000132693.12")
+
+# note that 'tissues' is a list-column
+x <- get_multi_tissue_eqtls(gencodeId = "ENSG00000132693.12",
+                            variantId = "chr1_159476920_T_C_b38")
+
+x$tissues[[1]]
 }
 }
 \seealso{
