### 2025-04-03 #684 @rmgpanw
Hi again, further to my [previous comment](https://github.com/ropensci/software-review/issues/684#issuecomment-2767540631) I have now added a new test file called `test-gtexr_api_match_checks.R`, which compares gtexr functions against the GTEx API specification. Commit [5c131ef](https://github.com/rmgpanw/gtexr/commit/5c131efe1cc36d9f2f8489f58fcd9b7db881e807)

I have fixed various issues based on the results and summarised these in `NEWS.md`.  This includes a new check in `gtex_query()` for API endpoints returning paginated responses, comparing the number of rows in the tibble output from gtexr with the number of data items in the raw API response. Commit [11e2c35](https://github.com/rmgpanw/gtexr/commit/11e2c35c3464237e68b3008c0d9302625dec393f)

I have also updated the R-CMD-check GitHub action to automatically run once a month. Commit [833ceac](https://github.com/rmgpanw/gtexr/commit/833ceac398985b71a7eb34c8185fa6f51f4cc9b9)
