### 2025-04-03 #684 @adamhsparks
> > [@mpadge](https://github.com/mpadge?rgh-link-date=2025-04-03T07%3A06%3A42.000Z), do I need to do something here to trigger the change in tags? I would have thought submitting the response with the changes based on the reviewers requests would have incremented it to the next tag.
>
> [@adamhsparks](https://github.com/adamhsparks?rgh-link-date=2025-04-03T07%3A06%3A42.000Z) I'm not sure what you mean there, sorry? Do you mean the issue labels, because that's now been moved to "4/reviews-in-awaiting-responses" as should be expected. Or do you mean something else?

Yes, do I need to set it to "5/Awaiting-reviewer(s)-response" manually? Or am I out of sync with the labels since we got the reviews back and @rmgpanw has made the reviewers' requested changes and we're waiting on them to approve the changes, aren't we?
