### 2025-04-03 #684 @mpadge
Aha, then yes you do indeed need to do that manually. Documented in last point in [this *Dev Guide* section](https://devguide.ropensci.org/softwarereview_editor.html#during-review) (just pasting that to reassure myself as well that we do indeed say that 😆 ).
