### 2025-04-15 #684 @adamhsparks
@rmgpanw, I just ran `devtools::check()` on your review branch version and have a failed test.

```r
── Failed tests ────────────────────────────────────────────────────────────────
Failure (test-download.R:14:3): `download()` returns tibble with expected colnam
es
names(result) (`actual`) not identical to c(...) (`expected`).

     actual            | expected
 [5] "rin"             | "rin"             [5]
 [6] "hasGTExImage"    | "hasGTExImage"    [6]
 [7] "concentration"   | "concentration"   [7]
 [8] "autolysisScore"  -
 [9] "analysisRelease" | "analysisRelease" [8]
[10] "genotype"        | "genotype"        [9]
[11] "hardyScale"      | "hardyScale"      [10]

[ FAIL 1 | WARN 0 | SKIP 0 | PASS 139 ]
```

Can you please check this?
