### 2025-04-15 #684 @rmgpanw
@adamhsparks I think you might have run this on the `main` branch, instead of the review branch `ropensci-software-review`?  I amended `test-download.R` in commit [d375585](https://github.com/rmgpanw/gtexr/commit/d3755853fca728f10421e7011fbcf2079974c5c3) and all tests seem to be passing for me currently.
