### 2025-04-15 #684 @ropensci-review-bot
Approved! Thanks @rmgpanw for submitting and @ernestguevarra, @zkamvar for your reviews! :grin:
