### 2025-04-16 #684 @rmgpanw
Fantastic, thank you so much! Thank you again for the impressive care effort you have all invested throughout the review process :)

@ernestguevarra and @zkamvar I would love to acknowledge you in the gtexr DESCRIPTION. Could you please confirm whether you are happy for me to do this. Of course, absolutely no worries if you prefer not.

@adamhsparks and @maurolepore, can I please also check whether any steps (in addition to those outlined) above are required for submission to [Journal of Open-Source Software](https://joss.theoj.org/) (JOSS)? The draft manuscript is located at inst/paper on the [ropensci-software-review](https://github.com/rmgpanw/gtexr/tree/ropensci-software-review/inst/paper) branch.
