### 2025-04-16 #684 @zkamvar
> I would love to acknowledge you in the gtexr DESCRIPTION. Could you please confirm whether you are happy for me to do this. Of course, absolutely no worries if you prefer not.

I am okay with the acknowledgement. Congratulations on the successful submission 🎉
