### 2025-04-16 #684 @adamhsparks
@rmgpanw, for submission to JOSS, we only perform the software review. JOSS then reviews the short paper you've prepared and accepts our review of the software.
